backend/pyproject.toml
[project]
name = "debate-panel-backend"
version = "0.1.0"
description = "DebatePanel V1 Backend"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "openai>=1.12.0",
    "pydantic>=2.6.0",
    "tiktoken>=0.5.0",
    "python-dotenv>=1.0.0",
    "crawl4ai>=0.4.0",
    "bubus>=0.1.0",
    "langdetect>=1.0.9",
    "duckduckgo-search>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0.0",
    "pytest-asyncio>=0.23.0",
    "httpx>=0.27.0",
]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"